pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "viceroybot"
description = 'A twitter bot for mimicking the writings of others'
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
keywords = []
authors = [
  { name = "frank1010111", email = "frmale@utexas.edu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "python-dotenv",
  "tweepy",
]
dynamic = ["version"]

[project.scripts]
viceroy = "viceroybot.viceroy:viceroy"

[project.urls]
Documentation = "https://github.com/frank1010111/viceroybot#readme"
Issues = "https://github.com/frank1010111/viceroybot/issues"
Source = "https://github.com/frank1010111/viceroybot"

[project.optional-dependencies]
dev = [
  "pre-commit"
]

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.version]
source = "code"
path = "viceroybot/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=viceroybot --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "viceroybot/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
